requirements.txt
streamlit
opencv-python
numpy
Pillow
streamlit-cropper
streamlit-image-coordinates
